% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perodicvalue.R
\name{makePeriodicValue}
\alias{makePeriodicValue}
\title{Make a periodically re-evaluated value-returning function.}
\usage{
makePeriodicValue(fun, period, ...)
}
\arguments{
\item{fun}{Function that generates a value.}

\item{period}{Period, inseconds, after which to refresh the value.}

\item{...}{Other arguments to be passed to the value-generating function.}
}
\value{
A function that returns a value, provided by the value-generating function,
and refreshes it after the provided refresh period.
}
\description{
Make a periodically re-evaluated value-returning function.
}
