% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchboard.R
\docType{class}
\name{addAction}
\alias{addAction}
\alias{delAction}
\alias{evaluateAll}
\alias{Switchboard-class}
\alias{Switchboard}
\alias{makeSwitchboard}
\title{Switchboard objects}
\usage{
addAction(.self, condition, action, name = NULL, endChain = F, ...)

delAction(.self, idx)

evaluateAll(.self)

makeSwitchboard()
}
\arguments{
\item{condition}{A logical statement, to be evaluated when the entire switchboard is evaluated.
If true, the associated action is executed.
The content of this argument remains unevaluated and is converted to a language object,
to be evaluated when the \code{evaluate} method is called.
However, content that's placed between \code{.()} is evaluated when running this function.}

\item{action}{A piece of code to be run when the associated if statement is evaluated and found to be true.
Content placed within \code{.()} is evaluated when running this function,
the rest remains unevaluated until the \code{evaluateAll} method is run.}

\item{name}{A name for this particular if-then statement. Can be left empty.}

\item{endChain}{Logical. Should this statement, if true, end the evaluation of any further statements?}

\item{idx}{Character or numeric. The to-be-deleted if-then statement. You can provide either its name or its number.}
}
\description{
switchboards are chains of if-then statements,
which can be very useful when long chains of such statements are needed,
such as when programming bots. Using switchboards allows users to divide parts of their chain
into different files or modules.
}
\section{Functions}{
\itemize{
\item \code{addAction}: add an action to a switchboard.

\item \code{delAction}: Delete an action from a switchboard.

\item \code{evaluateAll}: Evaluate all actions in a switchboard.
Evaluate all if-then statements in a switchboard.

\item \code{makeSwitchboard}: Make a switchboard object
}}

\section{Fields}{

\describe{
\item{\code{actions}}{A list of if-then statements.}
}}


