% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchboard.R
\docType{class}
\name{Switchboard-class}
\alias{Switchboard-class}
\alias{Switchboard}
\title{Switchboard objects}
\description{
switchboards are chains of if-then statements,
which can be very useful when long chains of such statements are needed,
such as when programming bots. Using switchboards allows users to divide parts of their chain
into different files or modules.
}
\section{Fields}{

\describe{
\item{\code{actions}}{A list of if-then statements.}
}}


